// web/src/shared/api/generated/model/paymentResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { BillingPeriod } from './billingPeriod';
import type { Money } from './money';
import type { PaymentStatus } from './paymentStatus';
import type { SubscriptionPlan } from './subscriptionPlan';

export interface PaymentResult {
  id: string;
  plan: SubscriptionPlan;
  period: BillingPeriod;
  amount: Money;
  status: PaymentStatus;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
}